refactor(pipeline): move pipeline and stream internals to separate modules
- Extracted internal pipeline execution coordination and fail-fast semantics
  into `cuprum/_pipeline_internals.py`.
- Added internal stream handling utilities for subprocess I/O in
  `cuprum/_streams.py`.
- Cleaned up `cuprum/sh.py` by importing relevant internals and removing
  duplicated code.
- Improved code modularity and separation of concerns for pipeline and
  streaming logic.
